MCPServer/pyproject.toml
[project]
name = "house-victoria-mcp"
version = "1.0.0"
description = "Advanced MCP server with persistent memory, complex tools, and specialized data banks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "piper-tts>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["house_victoria_mcp"]

[project.scripts]
house-victoria-mcp = "house_victoria_mcp.server:main"
